requirements.txt
Keras==2.2.4
pandas==0.24.2
xlwt==1.3.0
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
graphviz
pydot
numpy
matplotlib
python_docx
# pencv_python==4.1.1.26
xlrd
docx
Pillow
scikit_learn

